Route53 fix for zones with > 100 records
This uses methods from more recent (> 1.4.0) versions of fog to deal
with updating zones that have in excess of 100 records.  The AWS API
limits you to 100 records at a time when enumerating a zone.

The first change allows finding all of the matching records with
pagination.

The second is a much faster way to find whether or not a record
exists without having to deal with pagination.