bug fixes for conf_dir in mysql recipe
I'm not quite sure how this used to work on 11.04, but it is broken at
the moment. Files are being installed in /etc/mysql/mysql/conf.d, and
my.conf refers to the same directory. This update should fix it.

I've also changed from "platform" to "platform_family." Hope this
doesn't break anything for non-Ubuntu folks.