Ubuntu 12.04-related changes to Hue mysql setup
MySQL-python is complaining that it can't find a required library that
seems to have been updated in 12.04. I'm recompiling it for Hue to
make sure it's linked against the correct library.

In addition, it looks like an anonymous 'localhost' user without a
password is taking precedence over 'hue'@'%', which is making Hue's
login fail.  So I'm ensuring that the anonymous user is removed.