feature: adding nginx reverse proxy with basic auth flag to http listeners in srp
This is an update of the recent hotfix made for TV; we put nginx in front of http
listener apps on each http listener node, and we allow for enabling http basic
auth (username/password auth).

1. nginx has been added as a reverse proxy (default is from port 80 to port 8080)
   and in the future may be used for local load balancing as well (e.g. port 80
   to multiple [netty] apps on 8080, 8081, 8082)

2. nginx is deployed and configured via chef/ironfan/ironfan-plugins, and the http
   listeners are activated by tachyon as before

3. to enable basic_auth, just uncomment the appropriate lines in
   knife/credentials/knife-org.rb and flip the flag passed to the kafka_http_listener
   ironfan-plugin in your cluster definition

This is a coordinated update and involves modifications (on branches of the same feature/
review name) in the following repos:

infochimps-labs/ironfan-pantry
infochimps/enterprise-homebase
infochimps/enterprise-credentials
infochimps/srp-deploy
infochimps/ironfan-plugins

This has been tested against current master branches (with minor modifications, notably
using the stable branch of vayacondios) of the srp.

Squashed commit of the following:

commit 65e38cc798f018a3220ccd1783a2f0e8f4e80d75
Author: eric katerman <eric@infochimps.com>
Date:   Wed Feb 5 17:27:09 2014 -0800

    updating kafka cookbook to support new http listener features for srp

Conflicts:
	cookbooks/kafka/attributes/default.rb
	cookbooks/kafka/templates/default/kafka_contrib_proxy.conf.erb

Change-Id: Ic322b5cf0db9af9d6929e8e2c1ee5c5dc4488b08