fix: silverware no longer deletes old announcements
Previously, silverware used default attributes for announcements and
even took the additional step of deleting old announcements off of a
node. This made it impossible to run chef-client on a client and its
server simultaneously, even if both had previously announced: if node
B discovered components from node A and chef-client was run on both
simultaneously, the announcements could be wiped from A at the exact
moment that B needed to discover them.

Stale announcements will remain on a node, which should only cause
grief if a new machine has not already announced the correct service
in a realm: newer announcements will still be favored.

Change-Id: I36de88fbdce2975a85e187c9986b1d1394a601bf