fix: rotated logs exist where logs came from
1. I had hard-coded '/var/log/rotated' as the location where logs are
rotated. Logs need to be rotated to a directory that lives on the same
partition as the logs do: rotated logs for SRP components will then
live on /mnt/#{component}/ alongside unrotated file instead of the root
partition.

2. Logrotate requires that the 'olddir' location exist on the same
partition as the log being rotated. The location '/var/log/rotated' then
is not preferred if logs live on their own volume.

Change-Id: Iaf7fbef6dab2f74008406d191baf814b25d8c641